src/test/kotlin/DaySixTest.kt
import org.eruanno.DaySix
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.CsvSource

class DaySixTest {
    private val subject = DaySix()

    @ParameterizedTest
    @CsvSource(
        "'day-6-test', 41",
        "'day-6', -1"
    )
    fun testPartOne(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partOne(testFile))
    }

    @ParameterizedTest
    @CsvSource(
        "'day-6-test', 6",
        "'day-6', 1984"
    )
    fun testPartTwo(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partTwo(testFile))
    }
}
